@@ -6,10 +6,10 @@ import 'package:flutter_slidable/src/widgets/fractionnally_aligned_sized_box.dar
 import 'package:flutter_slidable/src/widgets/slidable_dismissal.dart';
 
 const double _kActionsExtentRatio = 0.25;
-const double _kFastThreshold = 2500.0;
+const double _kFastThreshold = 2500;
 const double _kDismissThreshold = 0.75;
-const Curve _kResizeTimeCurve = const Interval(0.4, 1.0, curve: Curves.ease);
-const Duration _kMovementDuration = const Duration(milliseconds: 200);
+const Curve _kResizeTimeCurve = Interval(0.4, 1, curve: Curves.ease);
+const Duration _kMovementDuration = Duration(milliseconds: 200);
 
 /// The rendering mode in which the [Slidable] is.
 enum SlidableRenderingMode {
@@ -39,17 +39,18 @@ enum SlideActionType {
 /// dismissed for the given [actionType].
 ///
 /// Used by [SlideToDismissDelegate.onDismissed].
-typedef void DismissSlideActionCallback(SlideActionType actionType);
+typedef DismissSlideActionCallback = void Function(SlideActionType? actionType);
 
 /// Signature for determining whether the widget will be dismissed for the
 /// given [actionType].
 ///
 /// Used by [SlideToDismissDelegate.onWillDismiss].
-typedef FutureOr<bool> SlideActionWillBeDismissed(SlideActionType actionType);
+typedef SlideActionWillBeDismissed = FutureOr<bool> Function(
+    SlideActionType? actionType);
 
 /// Signature for the builder callback used to create slide actions.
-typedef Widget SlideActionBuilder(BuildContext context, int index,
-    Animation<double> animation, SlidableRenderingMode step);
+typedef SlideActionBuilder = Widget Function(BuildContext context, int index,
+    Animation<double>? animation, SlidableRenderingMode step);
 
 /// A delegate that supplies slide actions.
 ///
@@ -71,7 +72,7 @@ abstract class SlideActionDelegate {
   /// Returns the child with the given index.
   ///
   /// Must not return null.
-  Widget build(BuildContext context, int index, Animation<double> animation,
+  Widget build(BuildContext context, int index, Animation<double>? animation,
       SlidableRenderingMode step);
 
   /// Returns the number of actions this delegate will build.
@@ -93,9 +94,9 @@ class SlideActionBuilderDelegate extends SlideActionDelegate {
   ///
   /// The [builder] must not be null. The [actionCount] argument must not be positive.
   const SlideActionBuilderDelegate({
-    @required this.builder,
-    @required this.actionCount,
-  }) : assert(actionCount != null && actionCount >= 0);
+    required this.builder,
+    required this.actionCount,
+  }) : assert(actionCount >= 0);
 
   /// Called to build slide actions.
   ///
@@ -104,10 +105,11 @@ class SlideActionBuilderDelegate extends SlideActionDelegate {
   final SlideActionBuilder builder;
 
   /// The total number of slide actions this delegate can provide.
+  @override
   final int actionCount;
 
   @override
-  Widget build(BuildContext context, int index, Animation<double> animation,
+  Widget build(BuildContext context, int index, Animation<double>? animation,
           SlidableRenderingMode step) =>
       builder(context, index, animation, step);
 }
@@ -128,27 +130,27 @@ class SlideActionListDelegate extends SlideActionDelegate {
   ///
   /// The [actions] argument must not be null.
   const SlideActionListDelegate({
-    @required this.actions,
+    required this.actions,
   });
 
   /// The slide actions.
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   /// The number of actions.
   @override
   int get actionCount => actions?.length ?? 0;
 
   @override
-  Widget build(BuildContext context, int index, Animation<double> animation,
+  Widget build(BuildContext context, int index, Animation<double>? animation,
           SlidableRenderingMode step) =>
-      actions[index];
+      actions![index];
 }
 
 class _SlidableScope extends InheritedWidget {
-  _SlidableScope({
-    Key key,
-    @required this.state,
-    @required Widget child,
+  const _SlidableScope({
+    Key? key,
+    required this.state,
+    required Widget child,
   }) : super(key: key, child: child);
 
   final SlidableState state;
@@ -159,25 +161,26 @@ class _SlidableScope extends InheritedWidget {
 
 /// The data used by a [Slidable].
 class SlidableData extends InheritedWidget {
-  SlidableData({
-    Key key,
-    @required this.actionType,
-    @required this.renderingMode,
-    @required this.totalActionsExtent,
-    @required this.dismissThreshold,
-    @required this.dismissible,
-    @required this.actionDelegate,
-    @required this.overallMoveAnimation,
-    @required this.actionsMoveAnimation,
-    @required this.dismissAnimation,
-    @required this.slidable,
-    @required this.actionExtentRatio,
-    @required this.direction,
-    @required Widget child,
+  /// Creates a [SlidableData].
+  const SlidableData({
+    Key? key,
+    required this.actionType,
+    required this.renderingMode,
+    required this.totalActionsExtent,
+    required this.dismissThreshold,
+    required this.dismissible,
+    required this.actionDelegate,
+    required this.overallMoveAnimation,
+    required this.actionsMoveAnimation,
+    required this.dismissAnimation,
+    required this.slidable,
+    required this.actionExtentRatio,
+    required this.direction,
+    required Widget child,
   }) : super(key: key, child: child);
 
   /// The type of slide action that is currently been showed by the [Slidable].
-  final SlideActionType actionType;
+  final SlideActionType? actionType;
 
   /// The rendering mode in which the [Slidable] is.
   final SlidableRenderingMode renderingMode;
@@ -193,16 +196,16 @@ class SlidableData extends InheritedWidget {
   final bool dismissible;
 
   /// The current actions that have to be shown.
-  final SlideActionDelegate actionDelegate;
+  final SlideActionDelegate? actionDelegate;
 
   /// Animation for the whole movement.
   final Animation<double> overallMoveAnimation;
 
   /// Animation for the actions.
-  final Animation<double> actionsMoveAnimation;
+  final Animation<double>? actionsMoveAnimation;
 
   /// Dismiss animation.
-  final Animation<double> dismissAnimation;
+  final Animation<double>? dismissAnimation;
 
   /// The slidable.
   final Slidable slidable;
@@ -233,22 +236,22 @@ class SlidableData extends InheritedWidget {
 
   /// If the [direction] is horizontal, returns the [totalActionsExtent]
   /// otherwise null.
-  double get actionPaneWidthFactor =>
+  double? get actionPaneWidthFactor =>
       directionIsXAxis ? totalActionsExtent : null;
 
   /// If the [direction] is vertical, returns the [totalActionsExtent]
   /// otherwise null.
-  double get actionPaneHeightFactor =>
+  double? get actionPaneHeightFactor =>
       directionIsXAxis ? null : totalActionsExtent;
 
   /// The data from the closest instance of this class that encloses the given context.
-  static SlidableData of(BuildContext context) {
+  static SlidableData? of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<SlidableData>();
   }
 
   /// Gets the the given offset related to the current direction.
   Offset createOffset(double value) {
-    return directionIsXAxis ? Offset(value, 0.0) : Offset(0.0, value);
+    return directionIsXAxis ? Offset(value, 0) : Offset(0, value);
   }
 
   /// Gets the maximum extent in the current direction.
@@ -258,9 +261,9 @@ class SlidableData extends InheritedWidget {
 
   /// Creates a positioned related to the current direction and showed actions.
   Positioned createPositioned({
-    Widget child,
-    double extent,
-    double position,
+    required Widget child,
+    double? extent,
+    double? position,
   }) {
     return Positioned(
       left: directionIsXAxis ? (showActions ? position : null) : 0.0,
@@ -275,9 +278,9 @@ class SlidableData extends InheritedWidget {
 
   /// Creates a [FractionallyAlignedSizedBox] related to the current direction and showed actions.
   FractionallyAlignedSizedBox createFractionallyAlignedSizedBox({
-    Widget child,
-    double extentFactor,
-    double positionFactor,
+    Widget? child,
+    double? extentFactor,
+    double? positionFactor,
   }) {
     return FractionallyAlignedSizedBox(
       leftFactor:
@@ -297,7 +300,7 @@ class SlidableData extends InheritedWidget {
   List<Widget> buildActions(BuildContext context) {
     return List.generate(
       actionCount,
-      (int index) => actionDelegate.build(
+      (int index) => actionDelegate!.build(
         context,
         index,
         actionsMoveAnimation,
@@ -334,22 +337,22 @@ class SlidableController {
   });
 
   /// Function called when the animation changed.
-  final ValueChanged<Animation<double>> onSlideAnimationChanged;
+  final ValueChanged<Animation<double>?>? onSlideAnimationChanged;
 
   /// Function called when the [Slidable] open status changed.
-  final ValueChanged<bool> onSlideIsOpenChanged;
+  final ValueChanged<bool?>? onSlideIsOpenChanged;
 
-  bool _isSlideOpen;
+  bool? _isSlideOpen;
 
-  Animation<double> _slideAnimation;
+  Animation<double>? _slideAnimation;
 
-  SlidableState _activeState;
+  SlidableState? _activeState;
 
   /// The state of the active [Slidable].
-  SlidableState get activeState => _activeState;
+  SlidableState? get activeState => _activeState;
 
   /// Changes the state of the active [Slidable].
-  set activeState(SlidableState value) {
+  set activeState(SlidableState? value) {
     _activeState?._flingAnimationController();
 
     _activeState = value;
@@ -360,19 +363,19 @@ class SlidableController {
         _slideAnimation?.addListener(_handleSlideIsOpenChanged);
         if (_slideAnimation == null) {
           _isSlideOpen = false;
-          onSlideIsOpenChanged(_isSlideOpen);
+          onSlideIsOpenChanged!(_isSlideOpen);
         }
       }
-      onSlideAnimationChanged(value?.overallMoveAnimation);
+      onSlideAnimationChanged!(value?.overallMoveAnimation);
     }
   }
 
   void _handleSlideIsOpenChanged() {
     if (onSlideIsOpenChanged != null && _slideAnimation != null) {
-      final bool isOpen = _slideAnimation.value != 0.0;
+      final bool isOpen = _slideAnimation!.value != 0.0;
       if (isOpen != _isSlideOpen) {
         _isSlideOpen = isOpen;
-        onSlideIsOpenChanged(_isSlideOpen);
+        onSlideIsOpenChanged!(_isSlideOpen);
       }
     }
   }
@@ -401,20 +404,20 @@ class Slidable extends StatefulWidget {
   /// state of the dismissed item. Using keys causes the widgets to sync
   /// according to their keys and avoids this pitfall.
   Slidable({
-    Key key,
-    @required Widget child,
-    @required Widget actionPane,
-    List<Widget> actions,
-    List<Widget> secondaryActions,
+    Key? key,
+    required Widget child,
+    required Widget actionPane,
+    List<Widget>? actions,
+    List<Widget>? secondaryActions,
     double showAllActionsThreshold = 0.5,
     double actionExtentRatio = _kActionsExtentRatio,
     Duration movementDuration = _kMovementDuration,
     Axis direction = Axis.horizontal,
     bool closeOnScroll = true,
     bool enabled = true,
-    SlidableDismissal dismissal,
-    SlidableController controller,
-    double fastThreshold,
+    SlidableDismissal? dismissal,
+    SlidableController? controller,
+    double? fastThreshold,
   }) : this.builder(
           key: key,
           child: child,
@@ -448,10 +451,10 @@ class Slidable extends StatefulWidget {
   /// which means the item after the dismissed item would be synced with the
   /// state of the dismissed item. Using keys causes the widgets to sync
   /// according to their keys and avoids this pitfall.
-  Slidable.builder({
-    Key key,
-    @required this.child,
-    @required this.actionPane,
+  const Slidable.builder({
+    Key? key,
+    required this.child,
+    required this.actionPane,
     this.actionDelegate,
     this.secondaryActionDelegate,
     this.showAllActionsThreshold = 0.5,
@@ -462,21 +465,11 @@ class Slidable extends StatefulWidget {
     this.enabled = true,
     this.dismissal,
     this.controller,
-    double fastThreshold,
-  })  : assert(actionPane != null),
-        assert(direction != null),
-        assert(
-            showAllActionsThreshold != null &&
-                showAllActionsThreshold >= .0 &&
-                showAllActionsThreshold <= 1.0,
+    double? fastThreshold,
+  })  : assert(showAllActionsThreshold >= .0 && showAllActionsThreshold <= 1.0,
             'showAllActionsThreshold must be between 0.0 and 1.0'),
-        assert(
-            actionExtentRatio != null &&
-                actionExtentRatio >= .0 &&
-                actionExtentRatio <= 1.0,
+        assert(actionExtentRatio >= .0 && actionExtentRatio <= 1.0,
             'actionExtentRatio must be between 0.0 and 1.0'),
-        assert(closeOnScroll != null),
-        assert(enabled != null),
         assert(dismissal == null || key != null,
             'a key must be provided if slideToDismissDelegate is set'),
         assert(fastThreshold == null || fastThreshold >= .0,
@@ -488,21 +481,21 @@ class Slidable extends StatefulWidget {
   final Widget child;
 
   /// The controller that tracks the active [Slidable] and keep only one open.
-  final SlidableController controller;
+  final SlidableController? controller;
 
   /// A delegate that builds slide actions that appears when the child has been dragged
   /// down or to the right.
-  final SlideActionDelegate actionDelegate;
+  final SlideActionDelegate? actionDelegate;
 
   /// A delegate that builds slide actions that appears when the child has been dragged
   /// up or to the left.
-  final SlideActionDelegate secondaryActionDelegate;
+  final SlideActionDelegate? secondaryActionDelegate;
 
   /// The action pane that controls how the slide actions are animated;
   final Widget actionPane;
 
   /// A delegate that controls how to dismiss the item.
-  final SlidableDismissal dismissal;
+  final SlidableDismissal? dismissal;
 
   /// Relative ratio between one slide action and the extent of the child.
   final double actionExtentRatio;
@@ -536,8 +529,8 @@ class Slidable extends StatefulWidget {
   final double fastThreshold;
 
   /// The state from the closest instance of this class that encloses the given context.
-  static SlidableState of(BuildContext context) {
-    final _SlidableScope scope =
+  static SlidableState? of(BuildContext context) {
+    final _SlidableScope? scope =
         context.dependOnInheritedWidgetOfExactType<_SlidableScope>();
     return scope?.state;
   }
@@ -567,51 +560,60 @@ class SlidableState extends State<Slidable>
     _dismissAnimation?.removeStatusListener(_handleShowAllActionsStatusChanged);
 
     _actionsMoveAnimation = CurvedAnimation(
-      parent: _overallMoveController,
-      curve: Interval(0.0, _totalActionsExtent),
+      parent: _overallMoveController!,
+      curve: Interval(0, _totalActionsExtent),
     )..addStatusListener(_handleShowAllActionsStatusChanged);
     _dismissAnimation = CurvedAnimation(
-      parent: _overallMoveController,
-      curve: Interval(_totalActionsExtent, 1.0),
+      parent: _overallMoveController!,
+      curve: Interval(_totalActionsExtent, 1),
     )..addStatusListener(_handleShowAllActionsStatusChanged);
   }
 
-  AnimationController _overallMoveController;
-  Animation<double> get overallMoveAnimation => _overallMoveController.view;
+  AnimationController? _overallMoveController;
 
-  Animation<double> _actionsMoveAnimation;
-  Animation<double> _dismissAnimation;
+  /// The overall animation.
+  Animation<double> get overallMoveAnimation => _overallMoveController!.view;
 
-  AnimationController _resizeController;
-  Animation<double> _resizeAnimation;
+  Animation<double>? _actionsMoveAnimation;
+  Animation<double>? _dismissAnimation;
 
-  double _dragExtent = 0.0;
+  AnimationController? _resizeController;
+  Animation<double>? _resizeAnimation;
+
+  double _dragExtent = 0;
 
   SlidableRenderingMode _renderingMode = SlidableRenderingMode.none;
+
+  /// Gets the active rendering mode.
   SlidableRenderingMode get renderingMode => _renderingMode;
 
-  ScrollPosition _scrollPosition;
+  ScrollPosition? _scrollPosition;
   bool _dragUnderway = false;
-  Size _sizePriorToCollapse;
+  Size? _sizePriorToCollapse;
   bool _dismissing = false;
 
-  SlideActionType _actionType = SlideActionType.primary;
-  SlideActionType get actionType => _actionType;
-  set actionType(SlideActionType value) {
+  SlideActionType? _actionType = SlideActionType.primary;
+
+  /// Gest the active type of action.
+  SlideActionType? get actionType => _actionType;
+
+  /// Sets the active type of action.
+  set actionType(SlideActionType? value) {
     _actionType = value;
     _initAnimations();
   }
 
   int get _actionCount => _actionDelegate?.actionCount ?? 0;
 
-  double get _totalActionsExtent => widget.actionExtentRatio * (_actionCount);
+  double get _totalActionsExtent => widget.actionExtentRatio * _actionCount;
 
   double get _dismissThreshold {
-    if (widget.dismissal == null)
+    if (widget.dismissal == null) {
       return _kDismissThreshold;
-    else
-      return widget.dismissal.dismissThresholds[actionType] ??
+    } else {
+      return widget.dismissal!.dismissThresholds[actionType!] ??
           _kDismissThreshold;
+    }
   }
 
   bool get _dismissible => widget.dismissal != null && _dismissThreshold < 1.0;
@@ -623,7 +625,7 @@ class SlidableState extends State<Slidable>
           _resizeController?.isAnimating == true);
 
   /// The current actions that have to be shown.
-  SlideActionDelegate get _actionDelegate =>
+  SlideActionDelegate? get _actionDelegate =>
       actionType == SlideActionType.primary
           ? widget.actionDelegate
           : widget.secondaryActionDelegate;
@@ -631,8 +633,8 @@ class SlidableState extends State<Slidable>
   bool get _directionIsXAxis => widget.direction == Axis.horizontal;
 
   double get _overallDragAxisExtent {
-    final Size size = context.size;
-    return _directionIsXAxis ? size.width : size.height;
+    final Size? size = context.size;
+    return _directionIsXAxis ? size!.width : size!.height;
   }
 
   double get _actionsDragAxisExtent {
@@ -659,20 +661,21 @@ class SlidableState extends State<Slidable>
   void _addScrollingNotifierListener() {
     if (widget.closeOnScroll) {
       _scrollPosition = Scrollable.of(context)?.position;
-      if (_scrollPosition != null)
-        _scrollPosition.isScrollingNotifier.addListener(_isScrollingListener);
+      if (_scrollPosition != null) {
+        _scrollPosition!.isScrollingNotifier.addListener(_isScrollingListener);
+      }
     }
   }
 
   void _removeScrollingNotifierListener() {
     if (_scrollPosition != null) {
-      _scrollPosition.isScrollingNotifier.removeListener(_isScrollingListener);
+      _scrollPosition!.isScrollingNotifier.removeListener(_isScrollingListener);
     }
   }
 
   @override
   void dispose() {
-    _overallMoveController.dispose();
+    _overallMoveController!.dispose();
     _resizeController?.dispose();
     _removeScrollingNotifierListener();
     widget.controller?._activeState = null;
@@ -682,7 +685,7 @@ class SlidableState extends State<Slidable>
   /// Opens the [Slidable].
   /// By default it's open the [SlideActionType.primary] action pane, but you
   /// can modify this by setting [actionType].
-  void open({SlideActionType actionType}) {
+  void open({SlideActionType? actionType}) {
     widget.controller?.activeState = this;
 
     if (actionType != null && _actionType != actionType) {
@@ -691,7 +694,7 @@ class SlidableState extends State<Slidable>
       });
     }
     if (_actionCount > 0) {
-      _overallMoveController.animateTo(
+      _overallMoveController!.animateTo(
         _totalActionsExtent,
         curve: Curves.easeIn,
         duration: widget.movementDuration,
@@ -701,7 +704,7 @@ class SlidableState extends State<Slidable>
 
   /// Closes this [Slidable].
   void close() {
-    if (!_overallMoveController.isDismissed) {
+    if (!_overallMoveController!.isDismissed) {
       if (widget.controller?.activeState == this) {
         widget.controller?.activeState = null;
       } else {
@@ -712,14 +715,14 @@ class SlidableState extends State<Slidable>
 
   void _flingAnimationController() {
     if (!_dismissing) {
-      _overallMoveController.fling(velocity: -1.0);
+      _overallMoveController!.fling(velocity: -1);
     }
   }
 
   /// Dismisses this [Slidable].
   /// By default it's dismiss by showing the [SlideActionType.primary] action pane, but you
   /// can modify this by setting [actionType].
-  void dismiss({SlideActionType actionType}) {
+  void dismiss({SlideActionType? actionType}) {
     if (_dismissible) {
       _dismissing = true;
       actionType ??= _actionType;
@@ -729,15 +732,17 @@ class SlidableState extends State<Slidable>
         });
       }
 
-      _overallMoveController.fling(velocity: 1.0);
+      _overallMoveController!.fling();
     }
   }
 
   void _isScrollingListener() {
-    if (!widget.closeOnScroll || _scrollPosition == null) return;
+    if (!widget.closeOnScroll || _scrollPosition == null) {
+      return;
+    }
 
     // When a scroll starts close this.
-    if (_scrollPosition.isScrollingNotifier.value) {
+    if (_scrollPosition!.isScrollingNotifier.value) {
       close();
     }
   }
@@ -745,33 +750,35 @@ class SlidableState extends State<Slidable>
   void _handleDragStart(DragStartDetails details) {
     _dragUnderway = true;
     widget.controller?.activeState = this;
-    _dragExtent =
-        _actionsMoveAnimation.value * _actionsDragAxisExtent * _dragExtent.sign;
-    if (_overallMoveController.isAnimating) {
-      _overallMoveController.stop();
+    _dragExtent = _actionsMoveAnimation!.value *
+        _actionsDragAxisExtent *
+        _dragExtent.sign;
+    if (_overallMoveController!.isAnimating) {
+      _overallMoveController!.stop();
     }
   }
 
   void _handleDragUpdate(DragUpdateDetails details) {
-    if (widget.controller != null && widget.controller.activeState != this) {
+    if (widget.controller != null && widget.controller!.activeState != this) {
       return;
     }
 
-    final double delta = details.primaryDelta;
+    final double delta = details.primaryDelta!;
     _dragExtent += delta;
     setState(() {
       actionType = _dragExtent.sign >= 0
           ? SlideActionType.primary
           : SlideActionType.secondary;
       if (_actionCount > 0) {
-        if (_dismissible && !widget.dismissal.dragDismissible) {
+        if (_dismissible && !widget.dismissal!.dragDismissible) {
           // If the widget is not dismissible by dragging, clamp drag result
           // so the widget doesn't slide past [_totalActionsExtent].
-          _overallMoveController.value =
+          _overallMoveController!.value =
               (_dragExtent.abs() / _overallDragAxisExtent)
-                  .clamp(0.0, _totalActionsExtent);
+                  .clamp(0.0, _totalActionsExtent)
+                  .toDouble();
         } else {
-          _overallMoveController.value =
+          _overallMoveController!.value =
               _dragExtent.abs() / _overallDragAxisExtent;
         }
       }
@@ -779,12 +786,12 @@ class SlidableState extends State<Slidable>
   }
 
   void _handleDragEnd(DragEndDetails details) {
-    if (widget.controller != null && widget.controller.activeState != this) {
+    if (widget.controller != null && widget.controller!.activeState != this) {
       return;
     }
 
     _dragUnderway = false;
-    final double velocity = details.primaryVelocity;
+    final double velocity = details.primaryVelocity!;
     final bool shouldOpen = velocity.sign == _dragExtent.sign;
     final bool fast = velocity.abs() > widget.fastThreshold;
 
@@ -795,7 +802,7 @@ class SlidableState extends State<Slidable>
       } else {
         open();
       }
-    } else if (_actionsMoveAnimation.value >= widget.showAllActionsThreshold ||
+    } else if (_actionsMoveAnimation!.value >= widget.showAllActionsThreshold ||
         (shouldOpen && fast)) {
       open();
     } else {
@@ -815,8 +822,8 @@ class SlidableState extends State<Slidable>
   }
 
   void _handleOverallPositionChanged() {
-    final double value = _overallMoveController.value;
-    if (value == _overallMoveController.lowerBound) {
+    final double value = _overallMoveController!.value;
+    if (value == _overallMoveController!.lowerBound) {
       _renderingMode = SlidableRenderingMode.none;
     } else if (value <= _totalActionsExtent) {
       _renderingMode = SlidableRenderingMode.slide;
@@ -827,13 +834,13 @@ class SlidableState extends State<Slidable>
     setState(() {});
   }
 
-  void _handleDismissStatusChanged(AnimationStatus status) async {
+  Future<void> _handleDismissStatusChanged(AnimationStatus status) async {
     if (_dismissible) {
       if (status == AnimationStatus.completed &&
-          _overallMoveController.value == _overallMoveController.upperBound &&
+          _overallMoveController!.value == _overallMoveController!.upperBound &&
           !_dragUnderway) {
-        if (widget.dismissal.onWillDismiss == null ||
-            await widget.dismissal.onWillDismiss(actionType)) {
+        if (widget.dismissal!.onWillDismiss == null ||
+            await widget.dismissal!.onWillDismiss!(actionType)) {
           _startResizeAnimation();
         } else {
           _dismissing = false;
@@ -850,19 +857,19 @@ class SlidableState extends State<Slidable>
 
   void _handleDismiss() {
     widget.controller?.activeState = null;
-    final SlidableDismissal dismissal = widget.dismissal;
+    final SlidableDismissal dismissal = widget.dismissal!;
     if (dismissal.onDismissed != null) {
       assert(actionType != null);
-      dismissal.onDismissed(actionType);
+      dismissal.onDismissed!(actionType);
     }
   }
 
   void _startResizeAnimation() {
     assert(_overallMoveController != null);
-    assert(_overallMoveController.isCompleted);
+    assert(_overallMoveController!.isCompleted);
     assert(_resizeController == null);
     assert(_sizePriorToCollapse == null);
-    final SlidableDismissal dismissal = widget.dismissal;
+    final SlidableDismissal dismissal = widget.dismissal!;
     if (dismissal.resizeDuration == null) {
       _handleDismiss();
     } else {
@@ -870,22 +877,22 @@ class SlidableState extends State<Slidable>
           AnimationController(duration: dismissal.resizeDuration, vsync: this)
             ..addListener(_handleResizeProgressChanged)
             ..addStatusListener((AnimationStatus status) => updateKeepAlive());
-      _resizeController.forward();
+      _resizeController!.forward();
       setState(() {
         _renderingMode = SlidableRenderingMode.resize;
         _sizePriorToCollapse = context.size;
-        _resizeAnimation = Tween<double>(begin: 1.0, end: 0.0).animate(
+        _resizeAnimation = Tween<double>(begin: 1, end: 0).animate(
             CurvedAnimation(
-                parent: _resizeController, curve: _kResizeTimeCurve));
+                parent: _resizeController!, curve: _kResizeTimeCurve));
       });
     }
   }
 
   void _handleResizeProgressChanged() {
-    if (_resizeController.isCompleted) {
+    if (_resizeController!.isCompleted) {
       _handleDismiss();
     } else {
-      widget.dismissal.onResize?.call();
+      widget.dismissal!.onResize?.call();
     }
   }
 
@@ -893,27 +900,27 @@ class SlidableState extends State<Slidable>
   Widget build(BuildContext context) {
     super.build(context); // See AutomaticKeepAliveClientMixin.
 
-    Widget content = widget.child;
+    Widget? content = widget.child;
 
     if (!(!widget.enabled ||
         ((widget.actionDelegate == null ||
-                widget.actionDelegate.actionCount == 0) &&
+                widget.actionDelegate!.actionCount == 0) &&
             (widget.secondaryActionDelegate == null ||
-                widget.secondaryActionDelegate.actionCount == 0)))) {
+                widget.secondaryActionDelegate!.actionCount == 0)))) {
       if (actionType == SlideActionType.primary &&
               widget.actionDelegate != null &&
-              widget.actionDelegate.actionCount > 0 ||
+              widget.actionDelegate!.actionCount > 0 ||
           actionType == SlideActionType.secondary &&
               widget.secondaryActionDelegate != null &&
-              widget.secondaryActionDelegate.actionCount > 0) {
+              widget.secondaryActionDelegate!.actionCount > 0) {
         if (_dismissible) {
           content = widget.dismissal;
 
           if (_resizeAnimation != null) {
             // we've been dragged aside, and are now resizing.
             assert(() {
-              if (_resizeAnimation.status != AnimationStatus.forward) {
-                assert(_resizeAnimation.status == AnimationStatus.completed);
+              if (_resizeAnimation!.status != AnimationStatus.forward) {
+                assert(_resizeAnimation!.status == AnimationStatus.completed);
                 throw FlutterError(
                     'A dismissed Slidable widget is still part of the tree.\n'
                     'Make sure to implement the onDismissed handler and to immediately remove the Slidable\n'
@@ -923,11 +930,11 @@ class SlidableState extends State<Slidable>
             }());
 
             content = SizeTransition(
-              sizeFactor: _resizeAnimation,
+              sizeFactor: _resizeAnimation!,
               axis: _directionIsXAxis ? Axis.vertical : Axis.horizontal,
               child: SizedBox(
-                width: _sizePriorToCollapse.width,
-                height: _sizePriorToCollapse.height,
+                width: _sizePriorToCollapse!.width,
+                height: _sizePriorToCollapse!.height,
                 child: content,
               ),
             );
